mm/page_alloc: use sched_clock() instead of jiffies to measure latency
sched_clock() (which is rdtsc() on x86) gives us more precise result
than jiffies.

Q: Why do we need greater accuracy?
A: Because if we target to, say, 10000 IOPS (per cpu) then
   1 ms memory allocation latency is too much and we need
   to achieve less alloc latency and thus measure it.

https://pmc.acronis.com/browse/VSTOR-19040


+++
mm/page_alloc: fix latency in tracepoint.

Since transition from jiffies to sched_clock(), mm_page_alloc
garbadge since we substract from jiffies nanoseconds.
Fix this.

mFixes: e9e7602565ab ("mm/page_alloc: use sched_clock() instead of jiffies to
measure latency")

+++
mm/page_alloc: fix latency in tracepoint - addon

Compilation warning fixed: use proper type.

mFixes: 7eff199a7a9e ("mm/page_alloc: fix latency in tracepoint.")


[VvS rh7-3.10.0-1062.4.1.el7 rebase]
 minor context change in __alloc_pages_nodemask()

+++ include/trace/events/kmem.h
@@ -194,7 +194,7 @@ TRACE_EVENT(mm_page_free_batched,
 TRACE_EVENT(mm_page_alloc,
 
 	TP_PROTO(struct page *page, unsigned int order,
-			gfp_t gfp_flags, int migratetype, unsigned int time),
+			gfp_t gfp_flags, int migratetype, u64 time),
 
 	TP_ARGS(page, order, gfp_flags, migratetype, time),
 
@@ -203,7 +203,7 @@ TRACE_EVENT(mm_page_alloc,
 		__field(	unsigned int,	order		)
 		__field(	gfp_t,		gfp_flags	)
 		__field(	int,		migratetype	)
-		__field(	unsigned int,	time		)
+		__field(	u64,		time		)
 	),
 
 	TP_fast_assign(
@@ -214,7 +214,7 @@ TRACE_EVENT(mm_page_alloc,
 		__entry->time		= time;
 	),
 
-	TP_printk("page=%p pfn=%lu order=%d migratetype=%d gfp_flags=%s lat=%d",
+	TP_printk("page=%p pfn=%lu order=%d migratetype=%d gfp_flags=%s lat=%llu",
 		__entry->page,
 		__entry->page ? page_to_pfn(__entry->page) : 0,
 		__entry->order,
